vlm-gui-fasthtml/requirements.txt
python-fasthtml>=0.1.0
uvicorn>=0.25.0
python-multipart>=0.0.6
httpx>=0.25.0
python-dotenv>=1.0.0 